fixed snackbar looks
