+++ lib/core/services/network_service.dart
@@ -50,7 +50,7 @@ class DioHttpServiceImpl implements DioHttpService {
       );
 
       return response.data;
-    } on DioError catch (e) {
+    } on DioError {
       rethrow;
     }
   }
